+++ alb-target-group.tf
@@ -1,5 +1,17 @@
+resource "random_string" "tg_suffix" {
+  length  = 4
+  upper   = false
+  special = false
+  keepers = {
+    name   = var.family
+    vpc_id = local.vpc_id
+    port   = var.container_port
+  }
+}
 resource "aws_lb_target_group" "alb" {
-  name                 = "${var.family}-tg"
+  # Excluding "name" field to allow for easier replacement when changing properties
+  name = "${var.family}-${random_string.tg_suffix.result}"
+  #name_prefix          = substr(var.family, 0, 6)
   port                 = var.container_port
   protocol             = "HTTP"
   target_type          = "ip"
@@ -16,6 +28,12 @@ resource "aws_lb_target_group" "alb" {
     unhealthy_threshold = 10
     matcher             = "200-399"
   }
+
+  tags = {
+    Name = "${var.family}-tg"
+  }
+
+  lifecycle { create_before_destroy = true }
 
   # apparently there is a bug with NLB stickiness right now
   # and this is the work around, not sure how it will affect ALBs